+++ apps/pages/views.py
@@ -6,4 +6,7 @@ from django.http import HttpResponse
 def index(request):
 
     # Page from the theme 
-    return render(request, 'pages/index.html')
+    context = {
+        'segment': 'dashboard'
+    }
+    return render(request, 'pages/index.html', context)