+++ CHANGELOG.md
@@ -46,7 +46,7 @@
 ### Improvements 
 
 - Bump Codebase - [Django Boilerplate Dashboard](https://github.com/app-generator/boilerplate-code-django-dashboard) **v1.0.4**
-    - Read properly the `.env` variables. Impacted file: **core/settings.py**.
+    - Read properly the `.env` variables. Impacted file: **config/settings.py**.
 
 ## [1.0.1] 2021-01-04
 ### Improvements